\name{AnalysisFunctions-package}
\alias{AnalysisFunctions-package}
\alias{AnalysisFunctions}
\docType{package}
\title{
\packageTitle{AnalysisFunctions}
}
\description{
\packageDescription{AnalysisFunctions}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{AnalysisFunctions}
\packageIndices{AnalysisFunctions}
Analysis functions for genomic data
}
\author{
\packageAuthor{AnalysisFunctions}

Maintainer: \packageMaintainer{AnalysisFunctions}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{

}
