% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enrichment_clusterProfiler.R
\name{enrichment.data4tyers}
\alias{enrichment.data4tyers}
\title{Run a universal enrichment (ORA) for a list of contrasts with cluterProfiler function "enricher"}
\usage{
enrichment.data4tyers(
  data4Tyers,
  contrast,
  gmt,
  collection_name = "",
  resultsDir = getwd(),
  translate_genes = FALSE,
  HOM_MouseHuman = "",
  minGSSize = 15,
  maxGSSize = 500,
  p.value = 0.05,
  p.adjust = 0.05,
  logFC = 0,
  plot_top = 50,
  plot.p.adjust = 0.05
)
}
\arguments{
\item{data4Tyers}{data4tyers dataframe. Must contain a Geneid column and
P.Value, adj.P.Val and logFC columns for each contrast (eg. "P.Value.LES_ACT.vs.LES_INACT")}

\item{contrast}{List with one vector of length 2 for each contrast.}

\item{gmt}{Data frame with gene sets. Can be obtained with function
clusterProfiler::read.gmt. Must have a column "term" and a column "gene".}

\item{collection_name}{Vector with the name of the collection. It will be
appended to the output directory/files name (eg. "c5.go.bp")}

\item{resultsDir}{Character vector with output results directory. Default is working directory.}

\item{translate_genes}{Wether to translate gene ids to homologues. If TRUE, a dataframe
with homologues must be provided (see HOM_MouseHuman), default is FALSE.}

\item{HOM_MouseHuman}{dataframe with HOM_MouseHumanSequence.txt. Must have the
following columns: HomoloGene.ID (common ID for homologues), Common.Organism.Name, Symbol}

\item{minGSSize}{minimal size of each geneSet for analyzing (default 15)}

\item{maxGSSize}{maximal size of genes annotated for testing (Default 500)}

\item{p.value}{include DE genes below this pvalue cutoff (default 0.05)}

\item{p.adjust}{include DE genes below this adjusted pvalue cutoff (default 0.05)}

\item{logFC}{logFC cutoff (default to 0)}

\item{plot_top}{maximum number of genesets to include in the dotplot}

\item{plot.p.adjust}{plot enrichment results below this adjusted pvalue cutoff (default 0.05)}
}
\value{
Returns a list with enrichemnt results for each contrast. The list is saved as a RData object
in the resultsDir directory, along with a folder for each contrast containing an
excel file with enrichment results.
}
\description{
Function that performs a gene-set enrichment analysis (ORA) using the function enricher
from clusterProfiler package. This function performs an hypergeometric test for each gene set,
for up an ddown-regulated genes separately.
}
\author{
Julia Perera Bel <jperera@imim.es>
}
