% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enrichment_clusterProfiler.R
\name{enrichment.geneList}
\alias{enrichment.geneList}
\title{Run a universal enrichment for a list of gens with cluterProfiler function "enricher"}
\usage{
enrichment.geneList(
  geneList,
  gmt,
  universe,
  collection_name = "",
  resultsDir = getwd(),
  specie = "human",
  HOM_MouseHuman = "",
  minGSSize = 15,
  maxGSSize = 500
)
}
\arguments{
\item{geneList}{list of vectors containing gene-lists. An enrichment will be 
performed for each position  of the list (ie. for each vector of genes).}

\item{gmt}{Data frame with gene sets. Can be obtained with function 
clusterProfiler::read.gmt. Must have a column "term" and a column "gene". 
The gene IDs must be in the same ofrmat as geneList (eg. SYMBOLS)}

\item{universe}{Background genes. Number of profiled genes}

\item{collection_name}{Vector with the name of the collection. It will be 
appended to the output directory/files name (eg. "c5.go.bp")}

\item{resultsDir}{Character vector with output results directory. Default is working directory.}

\item{specie}{Name of the specie. Accepts "human" and "mouse". If "mouse", a dataframe 
with homologues must be provided (see HOM_MouseHuman)}

\item{HOM_MouseHuman}{dataframe with HOM_MouseHumanSequence.txt. Must have the 
following columns: HomoloGene.ID (common ID for homologues), Common.Organism.Name, Symbol}

\item{minGSSize}{minimal size of each geneSet for analyzing (default 15)}

\item{maxGSSize}{maximal size of genes annotated for testing (Default 500)}
}
\value{
Returns a list with enrichemnt results for each contrast. The list is saved as a RData object 
in the resultsDir directory, along with a folder for each contrast containing an 
excel file with enrichment results.
}
\description{
Function that performs a gene-set enrichment analysis using the function enricher
from clusterProfiler package. This function performs an hypergeometric test for 
each gene set given a vector of genes.
}
\author{
Julia Perera Bel <jperera@imim.es>
}
