% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enrichment_clusterProfiler.R
\name{enrichment.plots}
\alias{enrichment.plots}
\title{Create plots summarizing enrichment results}
\usage{
enrichment.plots(
  enrichment,
  collection_name = "",
  resultsDir = getwd(),
  plot_top = 50,
  p.adjust = 0.05
)
}
\arguments{
\item{enrichment}{enrichment results obtained with enrichment.geneList or enrichment.data4tyers functions}

\item{collection_name}{Name of the collection to append to output files.}

\item{resultsDir}{Character vector with output results directory. Default is working directory.}

\item{plot_top}{maximum number of GSEA plots (default 50 to return all results)}

\item{p.adjust}{adjusted pvalue cutoff (default 0.05). Gene sets with p.adjust 
below this threshold will be included in all plots, unless maximum for visualization is reached.}
}
\value{
This function creates a folder for each contrast and generates
dotplots, gene-concept networks and enrichmentMAP for length(enrichment).
}
\description{
Function that creates dotplots, gene-concept networks 
and enrichmentMAP from clusterProfiler and enrichplot packages.
}
\author{
Julia Perera Bel <jperera@imim.es>
}
