% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.comparison.R
\name{excel.comparison}
\alias{excel.comparison}
\title{Perform comparisons between GSEA results}
\usage{
excel.comparison(
  pathinput,
  pathoutput,
  excelsUP,
  excelsDOWN,
  contrast,
  filename
)
}
\arguments{
\item{pathinput}{The pathway where excel files are}

\item{pathoutput}{The pathway for the excel file output}

\item{excelsUP}{The excels where functional pathways are upregulated (positive NES)}

\item{excelsDOWN}{The excels where functional pathways are downregulated (negative NES)}

\item{contrast}{The contrast used in functional analysis and it shows in excel files names}

\item{filename}{The name for the excel file output (without extension)}
}
\value{

}
\description{
Perform comparisons between GSEA results
}
