% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA_clusterProfiler.R
\name{GSEA.plots}
\alias{GSEA.plots}
\title{Create plots summarizing GSEA results}
\usage{
GSEA.plots(
  gsea,
  contrast,
  collection_name = "",
  resultsDir = getwd(),
  plot_top = 50,
  p.adjust = 0.05
)
}
\arguments{
\item{gsea}{GSEA results obtained with GSEA.run function}

\item{contrast}{List with one vector of length 2 for each contrast.}

\item{collection_name}{Name of the collection to append to output files.}

\item{resultsDir}{Character vector with output results directory. Default is working directory.}

\item{plot_top}{maximum number of GSEA plots (default 50 to return all results)}

\item{p.adjust}{adjusted pvalue cutoff (default 0.05). Gene sets with p.adjust 
below this threshold will be included in all plots, unless maximum for visualization is reached.}
}
\value{
This function creates a folder for each contrast and generates barplots, 
dotplots, GSEA plots, gene-concept networks and enrichmentMAP for both phenotypes.
}
\description{
Function that creates barplots, dotplots, GSEA plots, gene-concept networks 
and enrichmentMAP from clusterProfiler and enrichplot packages.
}
\author{
Julia Perera Bel <jperera@imim.es>
}
