% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA_clusterProfiler.R
\name{GSEA.run}
\alias{GSEA.run}
\title{Run GSEA for a list of contrasts with cluterProfiler wrapper for fgsea}
\usage{
GSEA.run(
  data4Tyers,
  contrast,
  gmt,
  resultsDir = getwd(),
  specie = "human",
  collection_name = "",
  minGSSize = 15,
  maxGSSize = 500,
  pvalueCutoff = 1,
  HOM_MouseHuman = ""
)
}
\arguments{
\item{data4Tyers}{data4tyers dataframe. Must contain a Geneid column and 
P.Value and logFC columns for each contrast (eg. "P.Value.LES_ACT.vs.LES_INACT")}

\item{contrast}{List with one vector of length 2 for each contrast.}

\item{gmt}{Data frame with gene sets. Can be obtained with function 
clusterProfiler::read.gmt. Must have a column "term" and a column "gene".}

\item{resultsDir}{Character vector with output results directory. Default is working directory.}

\item{specie}{Character vector with the name of the specie. If not "human", 
must provide an homology file (HOM_MouseHuman).}

\item{collection_name}{Name of the collection to append to output files.}

\item{minGSSize}{minimal size of each geneSet for analyzing (default 15)}

\item{maxGSSize}{maximal size of genes annotated for testing (Default 500)}

\item{pvalueCutoff}{adjusted pvalue cutoff (default 1 to return all results)}

\item{HOM_MouseHuman}{dataframe with HOM_MouseHumanSequence.txt. Must have 
the following columns: HomoloGene.ID (common ID for homologues), Common.Organism.Name, Symbol}
}
\value{
Returns a list with gsea results for each contrast. The list is saved An RData object 
in the resultsDir directory, along with a folder for each contrast containing an 
excel file with enrichment results.
}
\description{
Function that runs gene-set enrichment analysis (GSEA) using the function GSEA
from clusterProfiler package. This function calls fgsea package, a fast implementation 
of GSEA.
}
\author{
Julia Perera Bel <jperera@imim.es>
}
