\name{AnalysisFunctions-package}
\alias{AnalysisFunctions-package}
\alias{AnalysisFunctions}
\docType{package}
\title{
\packageTitle{AnalysisFunctions}
}
\description{
\packageDescription{AnalysisFunctions}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{AnalysisFunctions}
\packageIndices{AnalysisFunctions}
Analysis functions for genomic data
}
\author{
\packageAuthor{AnalysisFunctions}

Maintainer: \packageMaintainer{AnalysisFunctions}
}
\references{

}

\keyword{ package }
\seealso{

}
\examples{

}
